requirements.txt
# ==============================================================================
# REQUIREMENTS - TRANSCRIPTOR ÉLITE V4.0 (EDICIÓN FORENSE)
# Optimizado para Windows 10/11 con soporte NVIDIA CUDA 12.1
# ==============================================================================

# --- MOTOR DE TRANSCRIPCIÓN Y DIARIZACIÓN ---
whisperx==3.8.4
faster-whisper==1.2.1
pyannote-audio==4.0.4
openai-whisper==20250625

# --- NÚCLEO DE INTELIGENCIA (GPU CUDA 12.1) ---
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.8.2
pytorch-lightning==2.6.0

# --- PROCESAMIENTO DE AUDIO Y TEXTO ---
soundfile==0.13.1
python-docx==1.2.0
regex==2025.11.3
transformers==4.57.6
tokenizers==0.22.2

# --- INTERFAZ GRÁFICA Y UTILIDADES ---
Pillow==12.0.0
tqdm==4.67.1
requests==2.32.5
huggingface-hub==0.36.2
numpy==2.4.3
pandas==2.3.3

# --- HERRAMIENTAS DE DISTRIBUCIÓN ---
pyinstaller==6.17.0
auto-py-to-exe==2.48.1
